# src/backend/tests/contract/openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "FeatureFlagResponse": {
        "description": "Response for a single feature flag.",
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "feature": {
            "title": "Feature",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "feature",
          "enabled",
          "source"
        ],
        "title": "FeatureFlagResponse",
        "type": "object"
      },
      "FeatureFlagUpdate": {
        "description": "Request body for updating a feature flag.",
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          }
        },
        "required": [
          "enabled"
        ],
        "title": "FeatureFlagUpdate",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "AI-powered tech news aggregator API. Fetches articles from RSS feeds and videos from YouTube, processes with AI summarization, and serves personalized feeds.",
    "title": "blips-ai-news",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/flags": {
      "get": {
        "description": "Get status of all feature flags.\n\nReturns dict with each feature's current status and source (redis/env/default).",
        "operationId": "get_all_feature_flags_api_v1_admin_flags_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Admin-Key",
            "required": false,
            "schema": {
              "title": "X-Admin-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get All Feature Flags Api V1 Admin Flags Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get All Feature Flags",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/flags/{feature}": {
      "delete": {
        "description": "Reset a feature flag to its default value (removes Redis override).\n\nArgs:\n    feature: Feature name",
        "operationId": "reset_feature_flag_api_v1_admin_flags__feature__delete",
        "parameters": [
          {
            "in": "path",
            "name": "feature",
            "required": true,
            "schema": {
              "title": "Feature",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Admin-Key",
            "required": false,
            "schema": {
              "title": "X-Admin-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reset Feature Flag",
        "tags": [
          "admin"
        ]
      },
      "get": {
        "description": "Get status of a specific feature flag.\n\nArgs:\n    feature: Feature name (e.g., \"chat\", \"ingestion\")",
        "operationId": "get_feature_flag_api_v1_admin_flags__feature__get",
        "parameters": [
          {
            "in": "path",
            "name": "feature",
            "required": true,
            "schema": {
              "title": "Feature",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Admin-Key",
            "required": false,
            "schema": {
              "title": "X-Admin-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureFlagResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Feature Flag",
        "tags": [
          "admin"
        ]
      },
      "put": {
        "description": "Update a feature flag (stored in Redis for instant effect).\n\nArgs:\n    feature: Feature name (e.g., \"chat\", \"ingestion\")\n    update: New enabled status",
        "operationId": "update_feature_flag_api_v1_admin_flags__feature__put",
        "parameters": [
          {
            "in": "path",
            "name": "feature",
            "required": true,
            "schema": {
              "title": "Feature",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Admin-Key",
            "required": false,
            "schema": {
              "title": "X-Admin-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeatureFlagUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureFlagResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Feature Flag",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/maintenance/cleanup": {
      "post": {
        "description": "Manually trigger data retention cleanup.\n\nUses the same retention service as the daily scheduler.\nReturns counts of deleted rows per table.",
        "operationId": "manual_cleanup_api_v1_admin_maintenance_cleanup_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Admin-Key",
            "required": false,
            "schema": {
              "title": "X-Admin-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Manual Cleanup",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/maintenance/status": {
      "get": {
        "description": "Return last cleanup run info and current retention config.",
        "operationId": "cleanup_status_api_v1_admin_maintenance_status_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Admin-Key",
            "required": false,
            "schema": {
              "title": "X-Admin-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cleanup Status",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/stats/content": {
      "get": {
        "description": "Get content statistics for monitoring.",
        "operationId": "get_content_stats_api_v1_admin_stats_content_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Admin-Key",
            "required": false,
            "schema": {
              "title": "X-Admin-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Content Stats",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/trigger-fetch": {
      "post": {
        "description": "Manually trigger content ingestion + AI summarization.",
        "operationId": "trigger_fetch_api_v1_admin_trigger_fetch_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Admin-Key",
            "required": false,
            "schema": {
              "title": "X-Admin-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Trigger Fetch",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/trigger-summarize": {
      "post": {
        "description": "Manually trigger AI summarization for unprocessed content.",
        "operationId": "trigger_summarize_api_v1_admin_trigger_summarize_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Admin-Key",
            "required": false,
            "schema": {
              "title": "X-Admin-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Trigger Summarize",
        "tags": [
          "admin"
        ]
      }
    },
...
  }
}